refactor: move ActionQueue implementation to a separate module and guarantee action retrieval order
